requirements.txt
pandoc==2.4
plumbum==1.9.0
ply==3.11
pywin32==311
Flask
mysql-connector-python
openpyxl
